tools/local_stt/requirements.txt
faster-whisper>=1.0.3
ctranslate2>=4.4.0
torch>=2.2.0
torchaudio>=2.2.0
pyannote.audio>=3.1.1
